Python3 Compatibility
The package ConfigParser had its name changed to lower case on Python3. Python's version is verified with sys.version_info and manages the import.

Type unicode does not exist in Python3 and str is unicode by default. Python's version is verified and record unicode or str to the variable cast_class, used in line 29.